requirements.txt
streamlit
pandas
numpy
langchain-community
langchain-groq
langchain-core
python-dotenv
groq
pypdf
python-docx
beautifulsoup4
lxml
